chore(router): rename query type to indexable, add indexable types to router utils
