+++ libraries/grpc-sdk/src/interfaces/Indexable.ts
@@ -1,0 +1,7 @@
+export type Indexable = { [key: string]: any };
+export type Params = Indexable;
+export type Document = Indexable;
+export type Route = Indexable;
+export type Headers = Indexable;
+export type Context = Indexable;
+export type Query = Indexable;
