--- libraries/grpc-sdk/src/interfaces/Query.ts
@@ -1,7 +1,0 @@
-export type Query = { [key: string]: any };
-export type Params = Query;
-export type Document = Query;
-export type Route = Query;
-export type Headers = Query;
-export type Context = Query;
-export type Indexable = Query;
