+++ packages/router/src/interfaces/Indexable.ts
@@ -1,0 +1,2 @@
+export type Indexable = { [key: string]: any };
+export type Params = Indexable;
